Fixed datatype mismatch bug

+++ src/picc_temp_i2c.cpp
@@ -92,7 +92,7 @@ uint16_t hdc1010::readData(uint8_t pointer)
 	writeData(pointer) ;
 	uint8_t buf[2] ;
 	read ( i2cdevbus , buf , 2 ) ;
-	return (buf[0] << 8 | buf[1]) ;
+	return (((uint16_t)buf[0]) << 8 | buf[1]) ;
 }
 
 void hdc1010::writeData(uint8_t val)
